requirements_23.05.23.txt
appdirs==1.4.4
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16
deepspeed==0.9.2
docker-pycreds==0.4.0
einops==0.3.0
filelock==3.12.0
fsspec==2023.1.0
ftfy==6.1.1
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.31
hjson==3.1.0
huggingface-hub==0.0.8
idna==3.4
importlib-metadata==6.6.0
joblib==1.2.0
mpi4py==3.1.4
ninja==1.11.1
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.1
pandas==1.3.5
pathtools==0.1.2
Pillow==9.5.0
pkg_resources==0.0.0
protobuf==4.22.3
psutil==5.9.5
py-cpuinfo==9.0.0
pydantic==1.10.7
pydicom==2.3.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
sacremoses==0.0.53
sentry-sdk==1.21.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
timm==0.6.13
tokenizers==0.10.3
torch==1.13.1
torchtyping==0.1.4
torchvision==0.14.1
tqdm==4.65.0
transformers==4.6.0.dev0
typeguard==2.13.3
typing_extensions==4.5.0
urllib3==1.26.15
wandb==0.15.0
wcwidth==0.2.6
zipp==3.15.0
